# src/docs/swagger.yaml
definitions:
...
  dto.ReqUserSignIn:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.ReqUserSignUp:
    properties:
      about:
        type: string
      role:
        type: integer
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - name
    - password
    type: object
...
  dto.RespUser:
    properties:
      about:
        type: string
      role:
        type: integer
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  echo.HTTPError:
    properties:
      message: {}
    type: object
...
  models.User:
    properties:
      about:
        type: string
      role:
        type: integer
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
    type: object
  pkg.Response:
    properties:
      body: {}
    type: object
info:
  contact: {}
paths:
  /auth:
    get:
      description: check user auth
      produces:
      - application/json
      responses:
        "200":
          description: success auth
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                body:
                  $ref: '#/definitions/models.User'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Auth
      tags:
      - auth
...
  /logout:
    post:
      description: user logout
      produces:
      - application/json
      responses:
        "204":
          description: success logout, body is empty
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: invalid csrf
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Logout
      tags:
      - auth
...
  /signin:
    post:
      consumes:
      - application/json
      description: user sign in
      parameters:
      - description: user info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.ReqUserSignIn'
      produces:
      - application/json
      responses:
        "200":
          description: success sign in
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                body:
                  $ref: '#/definitions/dto.RespUser'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: invalid password
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: user doesn't exist
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: SignIn
      tags:
      - auth
  /signup:
    post:
      consumes:
      - application/json
      description: user sign up
      parameters:
      - description: user data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.ReqUserSignUp'
      produces:
      - application/json
      responses:
        "201":
          description: user created
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                body:
                  $ref: '#/definitions/dto.RespUser'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "409":
          description: email already exists
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: SignUp
      tags:
      - auth
...
swagger: "2.0"